+++ BTDB/StreamLayer/AbstractBufferedWriter.cs
@@ -5,6 +5,7 @@ using System.Net.Sockets;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using BTDB.Buffer;
+using Microsoft.Extensions.Primitives;
 
 namespace BTDB.StreamLayer
 {
@@ -483,6 +484,15 @@ namespace BTDB.StreamLayer
             WriteVUInt32((uint) value.Build + 1);
             if (value.Build == -1) return;
             WriteVUInt32((uint) value.Revision + 1);
+        }
+
+        public void WriteStringValues(StringValues value)
+        {
+            WriteVUInt32((uint)value.Count);
+            foreach (var s in value)
+            {
+                WriteString(s);
+            }
         }
     }
 }